# openapi.yaml
openapi: "3.1.0"
info:
  title: "Aegis API"
  version: "1.0.0"
  description: "Orchestrate Claude Code sessions via REST API. Aegis wraps Claude Code in ACP sessions\nand exposes everything through a unified API.\n\n**Base URL:** `http://localhost:9100`\n\n**Authentication:** All endpoints except `/v1/health` and `/health` require a bearer token.\nPass `Authorization: Bearer <your-token>` header, or set `AEGIS_AUTH_TOKEN` env var.\n\n**Session lifecycle:** `working` → `idle` → `permission_prompt` → `asking` → `stalled`\n\n**API versioning:** All `/v1/` responses include `X-Aegis-API-Version: 1`.\nLegacy unversioned paths (e.g. `/health`, `/sessions`) emit `Deprecation: true`,\n`Sunset`, and `X-API-Deprecated` headers. See `docs/api-versioning.md` for the\nfull versioning policy.\n"
servers:
  - url: "http://localhost:9100"
    description: "Local Aegis server"
security:
  - BearerAuth: []
paths:
...
  "/v1/sessions/{id}":
    get:
      operationId: "getSession"
      summary: "Get session details"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "200":
          description: "Session details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"




        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"



    delete:
      operationId: "killSession"
      summary: "Kill a session"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "204":
          description: "Session killed"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"
...
  "/v1/sessions/{id}/approve":
    post:
      operationId: "approvePermission"
      summary: "Approve a pending permission prompt"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "202":
          description: "Permission approved"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"

        "409":
          description: "No pending permission prompt"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  error:
                    type: "string"
  "/v1/sessions/{id}/reject":
    post:
      operationId: "rejectPermission"
      summary: "Reject a pending permission prompt"
      tags:
        - "Sessions"

      parameters:
        - $ref: "#/components/parameters/sessionId"


      responses:
        "202":
          description: "Permission rejected"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "404":
          $ref: "#/components/responses/NotFound"

        "409":
          description: "No pending permission prompt"
          content:
            application/json:
              schema:
                type: "object"
...
components:
  securitySchemes:
    BearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "API key"
  parameters:
    sessionId:
      name: "id"
      in: "path"
      required: true
      schema:
        type: "string"

      description: "Session ID"
...
  responses:
...
    Unauthorized:
      description: "Missing or invalid authentication token"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Unauthorized"
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              error:
                type: "string"
                example: "Not found"
  schemas:
...
    SessionInfo:
      type: "object"
      properties:
        id:
          type: "string"

        windowId:
          type: "string"

        windowName:
          type: "string"

        workDir:
          type: "string"

        claudeSessionId:
          type: "string"

        jsonlPath:
          type: "string"

        byteOffset:
          type: "integer"

        monitorOffset:
          type: "integer"

        status:
          type: "string"
          enum:
            - "idle"
            - "working"
            - "compacting"
            - "context_warning"
            - "waiting_for_input"
            - "permission_prompt"
            - "plan_mode"
            - "ask_question"
            - "bash_approval"
            - "settings"
            - "error"
            - "rate_limit"
            - "unknown"


        createdAt:
          type: "integer"

        lastActivity:
          type: "integer"

        stallThresholdMs:
          type: "integer"

        permissionStallMs:
          type: "integer"

        permissionMode:
          type: "string"

        settingsPatched:
          type: "boolean"

        hookSettingsFile:
          type: "string"

        hookSecret:
          type: "string"

        lastHookAt:
          type: "integer"

        activeSubagents:
          type: "array"
          items:
            type: "string"


        permissionPromptAt:
          type: "integer"

        permissionRespondedAt:
          type: "integer"

        lastHookReceivedAt:
          type: "integer"

        lastHookEventAt:
          type: "integer"

        model:
          type: "string"

        lastDeadAt:
          type: "integer"

        ccPid:
          type: "integer"

        parentId:
          type: "string"

        children:
          type: "array"
          items:
            type: "string"


        permissionPolicy:
          type: "array"
          items:
            type: "object"


        permissionProfile:
          type: "object"

        prd:
          type: "string"

        ownerKeyId:
          type: "string"
...